+++ idf_build_apps/main.py
@@ -52,7 +52,7 @@ def find_apps(
 
         rules = set()
         for _manifest_file in manifest_files:
-            LOGGER.info('Loading manifest file: %s', _manifest_file)
+            LOGGER.debug('Loading manifest file: %s', _manifest_file)
             rules.update(Manifest.from_file(_manifest_file).rules)
         manifest = Manifest(rules)
         App.MANIFEST = manifest
@@ -141,7 +141,7 @@ def build_apps(
         app.index = index
         app.verbose = build_verbose
 
-        LOGGER.debug('=> Building app %s: %s', index, repr(app))
+        LOGGER.info('=> Building app %s: %s', index, repr(app))
         is_built = False
         try:
             is_built = app.build(depends_on_components)